requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
Brotli==1.1.0
certifi==2024.2.2
click==8.1.7
distro==1.9.0
dj-database-url==2.1.0
Django==5.0.4
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
drf-spectacular==0.27.2
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
Markdown==3.6
numpy==1.26.4
openai==1.23.6
packaging==24.0
pandas==2.2.2
psycopg==3.1.18
psycopg-binary==3.1.18
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.0
rpds-py==0.18.0
ruff==0.3.7
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
uvicorn==0.29.0
whitenoise==6.6.0
